added angles

+++ core/src/com/haniel/Shooter/particles/BlueParticle.java
@@ -2,11 +2,11 @@ package com.haniel.Shooter.particles;
 
 public class BlueParticle extends Particle{
 
-	public BlueParticle(int x, int y, int life) {
+	public BlueParticle(double x, double y, int life) {
 		super(x, y, life);
 		this.texture = blueParticleTexture;
 	}
-	public BlueParticle(float x, float y, int life, double xa, double ya) {
+	public BlueParticle(double x, double y, int life, double xa, double ya) {
 		super(x, y, life, xa, ya);
 		this.texture = whiteParticleTexture;
 	}
