+++ core/src/com/haniel/Shooter/particles/BrownParticle.java
@@ -2,7 +2,7 @@ package com.haniel.Shooter.particles;
 
 public class BrownParticle  extends Particle{
 
-	public BrownParticle(int x, int y, int life) {
+	public BrownParticle(double x, double y, int life) {
 		super(x, y, life);
 		this.texture = brownParticleTexture;
 	}
