+++ core/src/com/haniel/Shooter/particles/OrangeParticle.java
@@ -2,11 +2,11 @@ package com.haniel.Shooter.particles;
 
 public class OrangeParticle extends Particle{
 
-	public OrangeParticle(float x, float y, int life) {
+	public OrangeParticle(double x, double y, int life) {
 		super(x, y, life);
 		this.texture = orangeParticleTexture;
 	}
-	public OrangeParticle(float x, float y, int life, double xa, double ya) {
+	public OrangeParticle(double x, double y, int life, double xa, double ya) {
 		super(x, y, life, xa, ya);
 		this.texture = orangeParticleTexture;
 	}
