+++ core/src/com/haniel/Shooter/particles/PurpleParticle.java
@@ -2,7 +2,7 @@ package com.haniel.Shooter.particles;
 
 public class PurpleParticle extends Particle{
 
-	public PurpleParticle(int x, int y, int life) {
+	public PurpleParticle(double x, double y, int life) {
 		super(x, y, life);
 		this.texture = purpleParticleTexture;
 	}
