+++ core/src/com/haniel/Shooter/particles/WhiteParticle.java
@@ -2,11 +2,11 @@ package com.haniel.Shooter.particles;
 
 public class WhiteParticle extends Particle{
 
-	public WhiteParticle(int x, int y, int life) {
+	public WhiteParticle(double x, double y, int life) {
 		super(x, y, life);
 		this.texture = whiteParticleTexture;
 	}
-	public WhiteParticle(float x, float y, int life, double xa, double ya) {
+	public WhiteParticle(double x, double y, int life, double xa, double ya) {
 		super(x, y, life, xa, ya);
 		this.texture = whiteParticleTexture;
 	}
